Progress job

+++ django/api/tasks.py
@@ -2,6 +2,7 @@ from mantistable.celery import app
 from api.models import Job
 
 import requests
+import json
 
 @app.task(name="test_task")
 def test_task(value, job_id):
@@ -14,19 +15,18 @@ def test_task(value, job_id):
 @app.task(name="rest_hook")
 def rest_hook(job_id):
     job = Job.objects.get(id=job_id)
-    callback = job.callback
+    job.progress["current"] += 1
+    job.save()
 
-    print(callback)
     try:
-        r = requests.post(callback, data={
-            'number': 12524,
-            'type': 'issue',
-            'action': 'test'
+        requests.post(job.callback, data={
+            'job_id': job.id,
+            'progress': json.dumps(job.progress),
         })
-        print(r.status_code, r.reason)
     except requests.exceptions.InvalidSchema as e:
         print(e)
     except requests.exceptions.ConnectionError as e:
         print(e)
 
-    job.delete()
+    if job.progress["total"] <= 0 or (job.progress["total"] > 0 and job.progress["current"] == job.progress["total"]):
+        job.delete()
